+++ engine/apps/api/views/shift_swap.py
@@ -46,6 +46,7 @@ class BaseShiftSwapViewSet(ModelViewSet):
 
     def _do_take(self, benefactor: User) -> dict:
         shift_swap = self.get_object()
+        prev_state = shift_swap.insight_logs_serialized
 
         try:
             shift_swap.take(benefactor)
@@ -58,6 +59,14 @@ class BaseShiftSwapViewSet(ModelViewSet):
 
         update_shift_swap_request_message.apply_async((ssr_pk,))
         notify_beneficiary_about_taken_shift_swap_request.apply_async((ssr_pk,))
+
+        write_resource_insight_log(
+            instance=shift_swap,
+            author=self.request.user,
+            event=EntityEvent.UPDATED,
+            prev_state=prev_state,
+            new_state=shift_swap.insight_logs_serialized,
+        )
 
         return ShiftSwapRequestSerializer(shift_swap).data