+++ engine/apps/slack/tests/conftest.py
@@ -1,4 +1,24 @@
+import typing
+
 import pytest
+from rest_framework import status
+from slack_sdk.web import SlackResponse
+
+
+def build_slack_response(
+    data: dict[str, typing.Any],
+    status_code: int = status.HTTP_200_OK,
+    headers: typing.Optional[dict[str, typing.Any]] = None,
+):
+    return SlackResponse(
+        client=None,
+        http_verb="POST",
+        api_url="test",
+        req_args={},
+        data=data,
+        headers=headers if headers else {},
+        status_code=status_code,
+    )
 
 
 @pytest.fixture