+++ engine/apps/slack/tests/test_scenario_steps/test_distribute_alerts.py
@@ -6,6 +6,7 @@ from apps.alerts.models import AlertGroup
 from apps.slack.client import SlackAPIException
 from apps.slack.models import SlackMessage
 from apps.slack.scenarios.scenario_step import ScenarioStep
+from apps.slack.tests.conftest import build_slack_response
 
 
 @pytest.mark.django_db
@@ -24,7 +25,9 @@ def test_restricted_action_error(
     step = SlackAlertShootingStep(slack_team_identity)
 
     with patch.object(step._slack_client, "api_call") as mock_slack_api_call:
-        mock_slack_api_call.side_effect = SlackAPIException(response={"error": "restricted_action"})
+        mock_slack_api_call.side_effect = SlackAPIException(
+            "error!", response=build_slack_response({"error": "restricted_action"})
+        )
         step._post_alert_group_to_slack(slack_team_identity, alert_group, alert, None, "channel-id", [])
 
     alert_group.refresh_from_db()